+++ RELEASE_NOTES.md
@@ -1,2 +1,5 @@
 v0.86.15
 
+- SVR2: update production enclave
+- SVRB: add new production enclave to `current` set
+